requirements.txt
browser-use==0.4.4
google-generativeai
python-dotenv